Extract Windows1257 to its own file
